MedianFilter/requirements.txt
numpy==1.14.0
Pillow==5.0.0

